+++ test/fixtures/linter/indentation_error_simple.sql
@@ -1,5 +1,5 @@
 SELECT
-   a.id -- 3 Spaces
-    a.name
+   a.id, -- 3 Spaces
+    a.name,
     a.training_spaces
 FROM tbl as a