+++ test/fixtures/linter/whitespace_errors.sql
@@ -1,0 +1,7 @@
+SELECT
+    a.id   , -- Comma with leading spaces
+    a.name,
+    a.training_spaces
+,  -- Comma on newline
+    a.tabs_alone_after_spaces
+FROM tbl as a