+++ test/linter_test.py
@@ -26,7 +26,7 @@ def test__linter__path_from_paths_dot():
     assert set(paths) >= set(['test\\fixtures\\lexer\\block_comment.sql', 'test\\fixtures\\lexer\\inline_comment.sql', 'test\\fixtures\\lexer\\basic.sql'])
 
 
-def test__linter__lint_file():
+def test__linter__lint_file_indentation():
     lntr = Linter()
     lnt = lntr.lint_path('test/fixtures/linter/indentation_errors.sql')
     # lets make an object of the codes, line numbers and positions of violations
@@ -36,7 +36,7 @@ def test__linter__lint_file():
             violations.append(elem)
     violations = [v.check_tuple() for v in violations]
     # Check we get the trialing whitespace violation
-    assert ('L001', 4, 22) in violations
+    assert ('L001', 4, 23) in violations
     # Check we get the mixed indentation errors
     assert ('L002', 3, 0) in violations
     assert ('L002', 4, 0) in violations
@@ -44,3 +44,17 @@ def test__linter__lint_file():
     assert ('L003', 3, 0) in violations
     # Check we get the mixed indentation errors between lines
     assert ('L004', 5, 0) in violations
+
+
+def test__linter__lint_file_whitespace():
+    lntr = Linter()
+    lnt = lntr.lint_path('test/fixtures/linter/whitespace_errors.sql')
+    # lets make an object of the codes, line numbers and positions of violations
+    violations = []
+    for file in lnt:
+        for elem in lnt[file]:
+            violations.append(elem)
+    violations = [v.check_tuple() for v in violations]
+    # Check we get comma whitespace errors
+    assert ('L005', 2, 11) in violations
+    assert ('L005', 5, 0) in violations